cleanup + refactor
