+++ frontend/package.json
@@ -29,6 +29,7 @@
 		"@radix-ui/react-tooltip": "^1.2.8",
 		"@replit/codemirror-vim": "6.2.1",
 		"@tailwindcss/vite": "^4.1.16",
+		"@tanstack/react-form": "^1.27.1",
 		"@tanstack/react-query": "^5.90.11",
 		"@tanstack/react-table": "^8.21.3",
 		"@uiw/codemirror-theme-aura": "4.22.2",
